# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaskFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [members]
      summary: List members of a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMember"
    post:
      operationId: inviteWorkspaceMember
      tags: [members]
      summary: Invite a member to a workspace by email
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceMemberInput"
      responses:
        "201":
          description: Created membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMember"
  /workspaces/{workspaceId}/members/{memberId}:
    delete:
      operationId: removeWorkspaceMember
      tags: [members]
      summary: Remove a member from a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "204":
          description: Removed
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    WorkspaceRole:
      type: string
      enum: [owner, admin, member]
    MemberStatus:
      type: string
      enum: [active, pending]
...
    WorkspaceMember:
      type: object
      required: [id, workspaceId, email, userId, role, status, createdAt]
      properties:
        id: { type: integer }
        workspaceId: { type: integer }
        email: { type: string }
        userId: { type: ["string", "null"] }
        role: { $ref: "#/components/schemas/WorkspaceRole" }
        status: { $ref: "#/components/schemas/MemberStatus" }
        createdAt: { type: string, format: date-time }
    WorkspaceMemberInput:
      type: object
      required: [email]
      properties:
        email: { type: string, minLength: 3 }
        role:
          type: string
          enum: [admin, member]
...