# lib/api-zod/src/generated/types/memberStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export type MemberStatus = typeof MemberStatus[keyof typeof MemberStatus];


export const MemberStatus = {
  active: 'active',
  pending: 'pending',
} as const;